fix(Sales): Justify the Notice & Disclaimer body on factsheet page 2
Extract the justified-text helpers into a shared scripts/factsheet/justify
module so page 2's disclaimer gets flush left/right edges instead of the
ragged mid-line wrapping from textwrap.fill.
